+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - new controller test
 
+### Changed
+- improvements to model classes
+
 ## 2019-10-10
 ### Fixed
 - controller code for the search book button